Revert "homepage cards missing"
This reverts commit d2ec343e4010bde859a68a1d99d29384546b7d0c.
