+++ package.json
@@ -10,7 +10,6 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
     "react-router-dom": "^6.4.2",
     "react-scripts": "^2.1.3",
     "react-transition-group": "^4.4.5",
