+++ src/components/AboutUs.js
@@ -1,0 +1,6 @@
+import React from "react";
+export default function AboutUs(){
+    return(
+        <p>AboutUs</p>
+    )
+}
