--- src/components/AboutUs/AboutUs.js
@@ -1,6 +1,0 @@
-import React from "react";
-export default function AboutUs(){
-    return(
-        <p>AboutUs</p>
-    )
-}
