+++ src/components/Appendix.js
@@ -1,0 +1,6 @@
+import React from "react";
+export default function Appendix(){
+    return(
+        <p>Appendix</p>
+    )
+}
