--- src/components/Appendix/Appendix.js
@@ -1,6 +1,0 @@
-import React from "react";
-export default function Appendix(){
-    return(
-        <p>Appendix</p>
-    )
-}
