--- src/components/CardList.js
@@ -1,5 +1,0 @@
-import React from "react";
-function CardList(){
-    return <p>hello</p>
-}
-export default CardList;
