+++ src/components/ContactUs.js
@@ -1,0 +1,6 @@
+import React from "react";
+export default function ContactUs(){
+    return(
+        <p>ContactUs</p>
+    )
+}
