--- src/components/ContactUs/ContactUs.js
@@ -1,6 +1,0 @@
-import React from "react";
-export default function ContactUs(){
-    return(
-        <p>ContactUs</p>
-    )
-}
