+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Footer() {
+    return(
+        <>
+            <div>Copyright ©2020 Pacific Agriscience Hong Kong Ltd. All rights reserved.</div>
+        </>
+    );
+}
+
+export default Footer;
