+++ src/components/Header.js
@@ -1,0 +1,8 @@
+// eslint-disable-next-line
+import React from 'react';
+
+export default function Header(){
+    return(
+        <></>
+    )
+}
