--- src/components/Header/Header.js
@@ -1,13 +1,0 @@
-// eslint-disable-next-line
-import React from 'react';
-import Helmet from 'react-helmet';
-
-export default function Header(){
-    return(
-        <div className='Header'>
-        <Helmet>
-            <title>Pacific Agriscience Hong Kong</title>
-        </Helmet>
-        </div>
-    )
-}
