+++ src/components/Home.js
@@ -1,0 +1,6 @@
+import React from "react";
+export default function Home(){
+    return(
+        <p>Home</p>
+    )
+}
