--- src/components/Home/index2.json
@@ -1,18 +1,0 @@
-[
-    {
-        "id": "aboutUs",
-        "image": "image/monkey.png"
-    },
-    {
-        "id": "productsAndServices",
-        "image": "image/monkey.png"
-    },
-    {
-        "id": "contactUs",
-        "image":"image/monkey.png"
-    },
-    {
-        "id": "mediaRoom",
-        "image":"image/monkey.png"
-    }
-]
