+++ src/components/MediaRoom.js
@@ -1,0 +1,6 @@
+import React from "react";
+export default function MediaRoom(){
+    return(
+        <p>MediaRoom</p>
+    )
+}
