--- src/components/MediaRoom/MediaRoom.js
@@ -1,6 +1,0 @@
-import React from "react";
-export default function MediaRoom(){
-    return(
-        <p>MediaRoom</p>
-    )
-}
