--- src/components/ProductsAndServices/ProductsAndServices.js
@@ -1,6 +1,0 @@
-import React from "react";
-export default function ProductsAndServices(){
-    return(
-        <p>ProductsAndServices</p>
-    )
-}
