+++ src/stylesheet.css
@@ -1,2 +1,51 @@
+/*Stylesheet.css*/
+		body {width:100%;
+			  background-image:url(3040791.jpg);
+			  margin: auto;
+			 }
+		div#top {background:linear-gradient(to bottom, #ccccff 0%, #ffffff 0%);
+			  	 padding: 10px; 
+		}
+		div#inner-top {max-width: 1600px;
+					   margin: auto;}
+		div #icons {text-align:right;}
+		#icons div {display: inline-block;
+					margin-top:10px;
+					padding:15px;
+					}
+		div#logo {max-width: 200px;
+			position: absolute;
+			margin-top: 5px;
+		}
+#returntotop {position: fixed;
+			  margin-left:94%;
+			  margin-top:32%}
+#returntotop img{width:80%;}
 
+div#toolbar {background:linear-gradient(to bottom, #ffffff 50%, #ccccff 100%);
+			 text-align: center;
+			border-top-style:solid;
+			border-bottom-style: solid;
+			border-width: thin;}
+
+
+.toolbarbuttons {display:inline-block;
+				font-family: calibri;
+				font-size: 20px;
+				color: black;
+				padding:20px;
+				padding-left:80px;
+				padding-right:80px;
+				font-weight: bold;}
+.toolbarbuttons:hover {background: linear-gradient(to bottom, #00ff00 0%, #009900 100%);} 
+
+
+		.footers{text-align: center;
+				font-weight: bold;
+				padding-top:20px;
+		 		padding-bottom: 30px;
+                background: linear-gradient(to bottom, #009933 0%, #336600 100%);
+				color: white;
+				bottom:0;
+		}
